#17 CML source file renamed from error.cml to errors.cml (plural)
