+++ cml-compiler/cml-language/src/main/java/cml/language/foundation/errors/Error.java
@@ -1,4 +1,4 @@
-package cml.language.foundation.error;
+package cml.language.foundation.errors;
 
 import cml.language.foundation.elements.ModelElement;
